Split out testutils package
