+++ java/compaction-job-execution/src/test/java/sleeper/compaction/jobexecution/testutils/CompactSortedFilesTestData.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.jobexecution;
+package sleeper.compaction.jobexecution.testutils;
 
 import com.facebook.collections.ByteArray;
 import org.apache.hadoop.fs.Path;
